Don't enqueue of add classes if setting doesn't allow field.
Best to not nq or add class unless that setting is currently allowed:
even if there is an URL/class value in postmeta b/c setting had been
on, a value added, then setting was turned off. If so, user wouldn’t
know value is there (unless setting was turned back on — i.e., to show
it in meta-box).